http_orm/orm_sqlalchemy/requirements.txt
sqlalchemy
python-decouple
psycopg2-binary